pyproject.toml
[project]
name = "funding-pipeline"
version = "0.1.0"
description = "Production-grade Indian startup funding intelligence pipeline — hierarchical scraping, LLM enrichment, schema-drift-aware, daily auto-run."
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "ryrk1020" }]

dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
    "playwright>=1.47",
    "pydantic>=2.8",
    "pydantic-settings>=2.4",
    "tenacity>=9.0",
    "loguru>=0.7",
    "pyyaml>=6.0",
    "python-dateutil>=2.9",
    "rapidfuzz>=3.9",
    "gspread>=6.1",
    "google-auth>=2.34",
    "typer>=0.12",
    "rich>=13.7",
    "tqdm>=4.66",
    "openai>=1.50",
    "python-dotenv>=1.0",
    "openpyxl>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-mock>=3.14",
    "ruff>=0.6",
    "mypy>=1.11",
    "respx>=0.21",
]

[project.scripts]
funding-pipeline = "pipeline.run:app"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["pipeline*", "sources*", "monitoring*", "config*"]
